Add cpack generator name as a required input to build-lnx-prod.sh

+++ scripts/build-lnx-prod.sh
@@ -5,10 +5,19 @@
 #
 #  Installs into /usr prefix directory under Linux.
 
+
+# Required input parameter check (used for the CPack generators)
+if [ "$1" == "" ]; then
+    echo "Usage: $0 <generator_names>"
+    echo 
+    echo "Example: $0 \"TGZ;DEB;RPM\""
+    exit 1
+fi
+
 # First build the application for Linux
 rm -rf build_prod
 cmake -GNinja -DCMAKE_INSTALL_PREFIX:PATH=/usr -DDOXYGEN:BOOL=FALSE -DPACKAGE=ON -DCMAKE_BUILD_TYPE=Release -B build_prod
 cmake --build ./build_prod --config Release 
 # Build packages
 cd build_prod
-cpack -C Release -G "TGZ;DEB;RPM"
+cpack -C Release -G "$1"
